requirements.txt
asgiref==3.10.0
Django==5.2.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-nested-routers==0.95.0
sqlparse==0.5.3
tzdata==2025.2
